requirements.txt
cryptography==41.0.7
Pillow==10.1.0
numpy==1.24.3
